Separate Stainless example into its own file
